requirements.txt
numpy==1.19.5
pandas==0.25.3
matplotlib==3.3.4
scikit-learn==0.22.2.post1
glob2==0.7
bunch==1.0.1
tensorflow==2.4.2
tensorflow-addons==0.13.0
torch==1.7.0
torchvision==0.8.1
open3d==0.8.0.0
opencv-python==4.4.0.46
psycopg2-binary==2.8.6
tqdm==4.52.0
ipywidgets==7.6.3
face_recognition==1.3.0
imgaug==0.4.0
xmltodict==0.12.0
pyntcloud==0.1.3
cgmzscore==2.0.4
wandb
azureml-sdk[notebooks]==1.21.0

flake8==3.9.2
pep8-naming==0.12.1
pytest==6.1.2
pytest-cov==2.10.1
pytest-azurepipelines==0.8.0
